codebook/requirements.txt
Django==1.11.2
Markdown==2.6.8
coverage==4.4.1
django-filter==1.0.4
djangorestframework==3.6.3
pytz==2017.2
